refactor(Dialog): useRemoteTrigger を RemoteDialogTrigger ディレクトリに移動
useRemoteTrigger の利用箇所が RemoteDialogTrigger/ 内のみのため、
Dialog/ 直下から RemoteDialogTrigger/ に移動する
